create Smart class
